added payment option

+++ pages/api/products/[id].js
@@ -5,6 +5,7 @@ export default async function handler(req, res) {
   const {
     method,
     query: { id },
+    cookies,
   } = req;
 
   await dbConnect();
@@ -19,6 +20,9 @@ export default async function handler(req, res) {
       }
       break;
     case "PUT":
+      if (!token || token !== process.env.TOKEN) {
+        return res.status(401).json("Not authorised");
+      }
       try {
         const product = await Product.create(req.body);
         res.status(201).json(product);
@@ -27,6 +31,9 @@ export default async function handler(req, res) {
       }
       break;
     case "DELETE":
+      if (!token || token !== process.env.TOKEN) {
+        return res.status(401).json("Not authorised");
+      }
       try {
         await Product.findByIdAndDelete(id);
         res.status(201).json("The product has been deleted!");
